# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: OnScreen API
  version: "2.2.0"
  description: |
    HTTP API for the OnScreen media server.

    **API stability** — frozen as of v2.2.0 (2026-05-09). Future v2.x
    minor releases may *add* endpoints, fields, or capability flags
    but never *remove* or *rename* them. Clients written against
    v2.2.0 keep working through the v2.x line. See
    [docs/server-lock.md](../server-lock.md) for the full posture
    and the v2.3 candidate list (ABR ladder, 2FA, Last.fm exporter,
    audio loudnorm).

    All JSON responses use a uniform envelope:
      - Success: `{ "data": <payload> }`
      - List:    `{ "data": [...], "meta": { "total": <int>, "cursor": "<string>" } }`
      - Error:   `{ "error": { "code": "<UPPER_SNAKE>", "message": "...", "request_id": "..." } }`

    The HLS playlist and segment endpoints under `/api/v1/transcode/sessions/{sid}/`
    are an intentional exception: HLS players cannot consume JSON errors, so those
    paths return raw `text/plain` errors with appropriate HTTP status codes.

    Auth is delivered as either a Paseto v4 bearer token (`Authorization: Bearer ...`)
    or the `os_access` httpOnly cookie set by `/auth/login`. Browser clients use the
    cookie; native and third-party clients use the bearer header. Asset routes
    (artwork, media stream, subtitles, downloads) additionally accept a
    `?token=<paseto>` query-string token so renderers that can't carry an
...
    authenticate.
  license:
    name: AGPL-3.0
servers:
...
security:
  - bearerAuth: []
  - cookieAuth: []
...
paths:
...
  /api/v1/admin/arr-services:
    get:
      summary: List configured arr services
      tags: [ArrServices]
      responses:
        '200':
          description: List of arr services. api_key is never returned; api_key_set indicates whether one is stored.
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/ArrService' }
                  meta: { $ref: '#/components/schemas/ListMeta' }
    post:
      summary: Add a new arr service
      tags: [ArrServices]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ArrServiceCreate' }
      responses:
        '201':
          description: Created.
          content:
            application/json:
              schema:
                type: object
                properties: { data: { $ref: '#/components/schemas/ArrService' } }
        '400': { description: Validation error. }
...
  /api/v1/admin/arr-services/{id}:
    parameters:
      - $ref: '#/components/parameters/UUID'
    get:
      summary: Get one arr service
      tags: [ArrServices]
      responses:
        '200':
          description: Service.
          content:
            application/json:
              schema:
                type: object
                properties: { data: { $ref: '#/components/schemas/ArrService' } }
        '404': { description: Not found. }
    patch:
      summary: Update an arr service
      description: |
        Partial update — every field is optional. Send `api_key` only when rotating
        the credential; an omitted or empty `api_key` keeps the existing value.
        `is_default` is intentionally not settable here — use /set-default.
      tags: [ArrServices]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ArrServiceUpdate' }
      responses:
        '200':
          description: Updated.
          content:
            application/json:
              schema:
                type: object
                properties: { data: { $ref: '#/components/schemas/ArrService' } }
        '404': { description: Not found. }
    delete:
      summary: Delete an arr service
      description: |
        In-flight requests bound to this service have their service_id set NULL via
        FK ON DELETE SET NULL — they survive but the admin must re-route them.
      tags: [ArrServices]
      responses:
        '204': { description: Deleted. }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: paseto-v4
    cookieAuth:
      type: apiKey
      in: cookie
      name: os_access
  parameters:
    UUID:
      in: path
      name: id
      required: true
      schema: { type: string, format: uuid }
...
  schemas:
...
    ListMeta:
      type: object
      properties:
        total:  { type: integer, format: int64, description: '`-1` when total is unknown (load-more pagination).' }
        cursor: { type: string }
    ArrService:
      type: object
      description: |
        A single Radarr/Sonarr instance. Multi-instance is supported per kind —
        is_default flags the one OnScreen will use when no per-request override
        is supplied.
      properties:
        id:                          { type: string, format: uuid }
        name:                        { type: string }
        kind:                        { type: string, enum: [radarr, sonarr] }
        base_url:                    { type: string }
        api_key_set:                 { type: boolean, description: api_key value is never returned; this flag indicates whether one is stored. }
        default_quality_profile_id:  { type: integer, nullable: true }
        default_root_folder:         { type: string,  nullable: true }
        default_tags:                { type: array, items: { type: integer } }
        minimum_availability:        { type: string, nullable: true, description: 'Radarr only: announced | inCinemas | released' }
        series_type:                 { type: string, nullable: true, description: 'Sonarr only: standard | daily | anime' }
        season_folder:               { type: boolean, nullable: true, description: Sonarr only. }
        language_profile_id:         { type: integer, nullable: true, description: Sonarr only (legacy v3). }
        is_default:                  { type: boolean }
        enabled:                     { type: boolean }
        created_at:                  { type: string, format: date-time }
        updated_at:                  { type: string, format: date-time }
    ArrServiceCreate:
      type: object
      required: [name, kind, base_url, api_key]
      properties:
        name:                        { type: string }
        kind:                        { type: string, enum: [radarr, sonarr] }
        base_url:                    { type: string }
        api_key:                     { type: string }
        default_quality_profile_id:  { type: integer, nullable: true }
        default_root_folder:         { type: string,  nullable: true }
        default_tags:                { type: array, items: { type: integer } }
        minimum_availability:        { type: string, nullable: true }
        series_type:                 { type: string, nullable: true }
        season_folder:               { type: boolean, nullable: true }
        language_profile_id:         { type: integer, nullable: true }
        is_default:                  { type: boolean, default: false }
        enabled:                     { type: boolean, default: true }
    ArrServiceUpdate:
      type: object
      description: |
        Every field is optional. Send `api_key` only when rotating the credential —
        an omitted or empty `api_key` keeps the existing value. is_default is not
        settable here; use POST /api/v1/admin/arr-services/{id}/set-default.
      properties:
        name:                        { type: string }
        base_url:                    { type: string }
        api_key:                     { type: string }
        default_quality_profile_id:  { type: integer, nullable: true }
        default_root_folder:         { type: string,  nullable: true }
        default_tags:                { type: array, items: { type: integer } }
        minimum_availability:        { type: string, nullable: true }
        series_type:                 { type: string, nullable: true }
        season_folder:               { type: boolean, nullable: true }
        language_profile_id:         { type: integer, nullable: true }
        enabled:                     { type: boolean }
...